Added custom hook useMobileFullScreen to check in App.js if the current device displaying the app is a mobile device to enforce portrait full screen mode
